use file size check instead of try except

+++ src/rule_utils/counts_to_tpm.py
@@ -2,6 +2,7 @@ import pandas as pd
 import numpy as np
 import argparse
 from pathlib import Path
+from os import path
 
 def get_args():
     parser = argparse.ArgumentParser()
@@ -76,13 +77,15 @@ def convert_dataframe(raw_data):
 def main():
     args = get_args()
     raw_data = pd.read_csv(args.readcounts)
-    try:
-        tpm = convert_dataframe(raw_data)
-        tpm.to_csv(args.output, sep="\t")
-    # If it is an empty file, just touch the output
-    except pd.errors.EmptyDataError:
+    
+    # If it is an empty file, just touch the output 
+    if (path.isfile(args.readcounts) and path.getsize(args.readcounts) > 0):
         print(f"{args.readcounts} is an EMPTY file. Creating EMPTY output at {args.output}")
-        Path(args.output).touch()    
+        Path(args.output).touch()  
+        return
+    
+    tpm = convert_dataframe(raw_data)
+    tpm.to_csv(args.output, sep="\t")
 
 
 if __name__=="__main__":
